+++ Code/Engine/TestFramework/Framework/SimpleTest.h
@@ -63,9 +63,9 @@ private:
 
 #define EZ_CREATE_SIMPLE_TEST(GroupName, TestName) \
   extern ezSimpleTestGroup EZ_CONCAT(g_SimpleTestGroup__,GroupName); \
-  static void ezSimpleTestFunction__##GroupName##TestName(); \
-  ezRegisterSimpleTestHelper ezRegisterSimpleTest__##GroupName##TestName(&EZ_CONCAT(g_SimpleTestGroup__,GroupName), EZ_STRINGIZE(TestName), ezSimpleTestFunction__##GroupName##TestName); \
-  static void ezSimpleTestFunction__##GroupName##TestName()
+  static void ezSimpleTestFunction__##GroupName##_##TestName(); \
+  ezRegisterSimpleTestHelper ezRegisterSimpleTest__##GroupName##TestName(&EZ_CONCAT(g_SimpleTestGroup__,GroupName), EZ_STRINGIZE(TestName), ezSimpleTestFunction__##GroupName##_##TestName); \
+  static void ezSimpleTestFunction__##GroupName##_##TestName()